registerCarouselのhooksを分離
